requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
requests>=2.31.0
pydantic>=2.6.0
streamlit>=1.30.0
python-dotenv>=1.0.0
sentence-transformers>=2.3.1
faiss-cpu>=1.7.4
func-timeout>=4.3.5
numpy>=1.26.0
google-generativeai>=0.3.0